backend/requirements_cv.txt
# AROHAN CV Backend Requirements

flask==3.0.0
flask-cors==4.0.0
opencv-contrib-python   # DNN + Haar cascade face detection
numpy
requests
Pillow
python-dotenv

# Primary recognizer — FaceNet based (requires cmake + dlib)
# Install with: pip install face-recognition
# If this fails, the backend auto-falls back to LBP+HOG (no extra deps needed)
face-recognition

# Optional: Firebase snapshot upload + alert sync
firebase-admin
